Make the Coronavirus Helper methods a module - and include those in the relevant specs - remove unecessary require_relatives
